28. Find the Index of the First Occurrence in a String
